Moved calculating functions to own script file
